api/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
psycopg2-binary
pydantic-settings
